refactor: move the BaseCursor class to its own module
